frontend/requirements_fe.txt
requests>=2.31.0
Flask>=2.3.3
Pillow==10.0.0
